Photo Uploading v0.2 / Bug Fixes
DAAAYUM, IT'S MUCH HARDER THAN I THOUGHT IT WOULD BE!!!!

- Now we have working "Manage Photos" page, that has a list of rooms and
images for everyroom

Trevor, I am sorry, I had to change some of your code in order to get it
to work. Also, fixed several errors. Please let me know if something
doesn't work the way it's supposed to.

+++ src/main/java/hotelmaster/gallery/PhotoRowMapper.java
@@ -6,7 +6,6 @@
 package hotelmaster.gallery;
 
 import hotelmaster.Photo;
-import hotelmaster.account.Account;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import org.springframework.jdbc.core.RowMapper;
@@ -21,8 +20,8 @@ public class PhotoRowMapper implements RowMapper {
     public Object mapRow(ResultSet rs, int rowNum) throws SQLException {
         Photo photo = new Photo();
         
-        photo.setImageID(rs.getLong("image_id"));
-        photo.setRoomID(rs.getLong("room_id"));
+        photo.setImageID(rs.getInt("image_id"));
+        photo.setRoomID(rs.getInt("room_id"));
         photo.setImage(rs.getBytes("image"));
         photo.setTitle(rs.getString("title"));
